git commit -m"refactor(folders):  implement LFA feedback
- remove unnecessary abstracted functions
 [Finishes  #159423414]
