+++ server/controllers/diaryController/addDiaryEntryController.js
@@ -1,7 +1,8 @@
 import postDiaryEntryValidator from '../../helpers/diaryIndexValidator';
-import addDiaryEntryDatabaseCall from './addDiaryEntryQuery';
 import authenticateToken from '../../middlewares/authenticateToken';
 import doesUserExist from '../../middlewares/doesUserExist';
+import isPrivacyFieldEmpty from '../../middlewares/isPrivacyEmpty';
+import db from '../../models/db';
 /**
  * @class diarycontroller
  */
@@ -13,17 +14,23 @@ class diarycontroller {
    * @returns {JSON} returns a JSON object
    */
   static addEntry(req, res) {
-    // check if there are any validation errors
     const { errors, isValid } = postDiaryEntryValidator(req.body, false);
-    /** isEditing argument set to false, because it's the POST route
-    POST and PUT API routes share the same validator
-    */
     if (isValid) {
-      // validate the token
       authenticateToken(req, res, (decodedToken) => {
         doesUserExist(res, decodedToken.userId, (userId) => {
-          // send the results of the database call if there are no validation errors
-          addDiaryEntryDatabaseCall(req, res, userId);
+          req.body.privacy = isPrivacyFieldEmpty(req.body);
+          const text = 'INSERT INTO entries(title, description, privacy, userid) VALUES ($1, $2, $3, $4) RETURNING \n'
+          + 'id, title, description, privacy';
+          const params = [req.body.title, req.body.description, req.body.privacy, userId];
+          db(text, params, (err, result) => {
+            if (err) {
+              return res.status(500).json({ error: err.stack });
+            }
+            res.status(201).json({
+              message: 'You have successfully posted the diary entry',
+              data: result.rows[0]
+            });
+          });
         });
       });
     } else {
