--- server/controllers/userController/verifyUserPassword.js
@@ -1,13 +1,0 @@
-import bcrypt from 'bcrypt';
-
-const verifyPassword = (password, hash, callback) => {
-  bcrypt.compare(password, hash)
-    .then((isMatch) => {
-      callback(isMatch);
-    })
-    .catch((err) => {
-      throw err;
-    });
-};
-
-export default verifyPassword;
